delete SatscribeDescriptionRepository in favor of ConversationRepository
